rename game2 to game
